+++ sdk/resourcemanager/chaos/armchaos/models.go
@@ -175,14 +175,23 @@ type CapabilityTypeListResult struct {
 
 // CapabilityTypeProperties - Model that represents the Capability Type properties model.
 type CapabilityTypeProperties struct {
+	// Runtime properties of this Capability Type.
+	RuntimeProperties *CapabilityTypePropertiesRuntimeProperties
+
 	// READ-ONLY; Localized string of the description.
 	Description *string
 
 	// READ-ONLY; Localized string of the display name.
 	DisplayName *string
 
+	// READ-ONLY; String of the kind of this Capability Type.
+	Kind *string
+
 	// READ-ONLY; URL to retrieve JSON schema of the Capability Type parameters.
 	ParametersSchema *string
+
+	// READ-ONLY; Array of control and data plane actions necessary to execute capability type.
+	PermissionsNecessary *CapabilityTypePropertiesPermissionsNecessary
 
 	// READ-ONLY; String of the Publisher that this Capability Type extends.
 	Publisher *string
@@ -194,6 +203,22 @@ type CapabilityTypeProperties struct {
 	Urn *string
 }
 
+// CapabilityTypePropertiesPermissionsNecessary - Array of control and data plane actions necessary to execute capability
+// type.
+type CapabilityTypePropertiesPermissionsNecessary struct {
+	// Control plane actions necessary to execute capability type.
+	Actions []*string
+
+	// Control plane actions necessary to execute capability type.
+	DataActions []*string
+}
+
+// CapabilityTypePropertiesRuntimeProperties - Runtime properties of this Capability Type.
+type CapabilityTypePropertiesRuntimeProperties struct {
+	// READ-ONLY; String of the kind of the resource's action type (continuous or discrete).
+	Kind *string
+}
+
 // CapabilityTypesClientGetOptions contains the optional parameters for the CapabilityTypesClient.Get method.
 type CapabilityTypesClientGetOptions struct {
 	// placeholder for future optional parameters
@@ -203,6 +228,15 @@ type CapabilityTypesClientGetOptions struct {
 type CapabilityTypesClientListOptions struct {
 	// String that sets the continuation token.
 	ContinuationToken *string
+}
+
+// ComponentsEwb5TmSchemasUserassignedidentitiesAdditionalproperties - The list of user identities associated with the experiment.
+type ComponentsEwb5TmSchemasUserassignedidentitiesAdditionalproperties struct {
+	// READ-ONLY; The client id of user assigned identity.
+	ClientID *string
+
+	// READ-ONLY; The principal id of user assigned identity.
+	PrincipalID *string
 }
 
 // ContinuousAction - Model that represents a continuous action.
@@ -441,7 +475,7 @@ type ExperimentListResult struct {
 // ExperimentProperties - Model that represents the Experiment properties model.
 type ExperimentProperties struct {
 	// REQUIRED; List of selectors.
-	Selectors []*Selector
+	Selectors []SelectorClassification
 
 	// REQUIRED; List of steps.
 	Steps []*Step
@@ -495,17 +529,20 @@ type ExperimentStatusProperties struct {
 	Status *string
 }
 
-// ExperimentsClientBeginCancelOptions contains the optional parameters for the ExperimentsClient.BeginCancel method.
-type ExperimentsClientBeginCancelOptions struct {
-	// Resumes the LRO from the provided token.
-	ResumeToken string
+// ExperimentUpdate - Describes an experiment update.
+type ExperimentUpdate struct {
+	// The identity of the experiment resource.
+	Identity *ResourceIdentity
 }
 
-// ExperimentsClientBeginCreateOrUpdateOptions contains the optional parameters for the ExperimentsClient.BeginCreateOrUpdate
-// method.
-type ExperimentsClientBeginCreateOrUpdateOptions struct {
-	// Resumes the LRO from the provided token.
-	ResumeToken string
+// ExperimentsClientCancelOptions contains the optional parameters for the ExperimentsClient.Cancel method.
+type ExperimentsClientCancelOptions struct {
+	// placeholder for future optional parameters
+}
+
+// ExperimentsClientCreateOrUpdateOptions contains the optional parameters for the ExperimentsClient.CreateOrUpdate method.
+type ExperimentsClientCreateOrUpdateOptions struct {
+	// placeholder for future optional parameters
 }
 
 // ExperimentsClientDeleteOptions contains the optional parameters for the ExperimentsClient.Delete method.
@@ -564,6 +601,29 @@ type ExperimentsClientStartOptions struct {
 	// placeholder for future optional parameters
 }
 
+// ExperimentsClientUpdateOptions contains the optional parameters for the ExperimentsClient.Update method.
+type ExperimentsClientUpdateOptions struct {
+	// placeholder for future optional parameters
+}
+
+// FilterClassification provides polymorphic access to related types.
+// Call the interface's GetFilter() method to access the common type.
+// Use a type switch to determine the concrete type.  The possible types are:
+// - *Filter, *SimpleFilter
+type FilterClassification interface {
+	// GetFilter returns the Filter content of the underlying type.
+	GetFilter() *Filter
+}
+
+// Filter - Model that represents available filter types that can be applied to a targets list.
+type Filter struct {
+	// REQUIRED; Enum that discriminates between filter types. Currently only Simple type is supported.
+	Type *FilterType
+}
+
+// GetFilter implements the FilterClassification interface for type Filter.
+func (f *Filter) GetFilter() *Filter { return f }
+
 // KeyValuePair - A map to describe the settings of an action.
 type KeyValuePair struct {
 	// REQUIRED; The name of the setting for the action.
@@ -571,6 +631,34 @@ type KeyValuePair struct {
 
 	// REQUIRED; The value of the setting for the action.
 	Value *string
+}
+
+// ListSelector - Model that represents a list selector.
+type ListSelector struct {
+	// REQUIRED; String of the selector ID.
+	ID *string
+
+	// REQUIRED; List of Target references.
+	Targets []*TargetReference
+
+	// REQUIRED; Enum of the selector type.
+	Type *SelectorType
+
+	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
+	AdditionalProperties map[string]any
+
+	// Model that represents available filter types that can be applied to a targets list.
+	Filter FilterClassification
+}
+
+// GetSelector implements the SelectorClassification interface for type ListSelector.
+func (l *ListSelector) GetSelector() *Selector {
+	return &Selector{
+		Type:                 l.Type,
+		ID:                   l.ID,
+		Filter:               l.Filter,
+		AdditionalProperties: l.AdditionalProperties,
+	}
 }
 
 // Operation - Details of a REST API operation, returned from the Resource Provider Operations API
@@ -627,6 +715,37 @@ type OperationsClientListAllOptions struct {
 	// placeholder for future optional parameters
 }
 
+// QuerySelector - Model that represents a query selector.
+type QuerySelector struct {
+	// REQUIRED; String of the selector ID.
+	ID *string
+
+	// REQUIRED; Azure Resource Graph (ARG) Query Language query for target resources.
+	QueryString *string
+
+	// REQUIRED; Subscription id list to scope resource query.
+	SubscriptionIDs []*string
+
+	// REQUIRED; Enum of the selector type.
+	Type *SelectorType
+
+	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
+	AdditionalProperties map[string]any
+
+	// Model that represents available filter types that can be applied to a targets list.
+	Filter FilterClassification
+}
+
+// GetSelector implements the SelectorClassification interface for type QuerySelector.
+func (q *QuerySelector) GetSelector() *Selector {
+	return &Selector{
+		Type:                 q.Type,
+		ID:                   q.ID,
+		Filter:               q.Filter,
+		AdditionalProperties: q.AdditionalProperties,
+	}
+}
+
 // Resource - Common fields that are returned in the response for all Azure Resource Manager resources
 type Resource struct {
 	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
@@ -639,10 +758,15 @@ type Resource struct {
 	Type *string
 }
 
-// ResourceIdentity - The managed identity of a resource.
+// ResourceIdentity - The identity of a resource.
 type ResourceIdentity struct {
 	// REQUIRED; String of the resource identity type.
 	Type *ResourceIdentityType
+
+	// The list of user identities associated with the experiment. The user identity dictionary key references will be ARM resource
+	// ids in the form:
+	// '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ManagedIdentity/userAssignedIdentities/{identityName}'.
+	UserAssignedIdentities map[string]*ComponentsEwb5TmSchemasUserassignedidentitiesAdditionalproperties
 
 	// READ-ONLY; GUID that represents the principal ID of this resource identity.
 	PrincipalID *string
@@ -651,16 +775,53 @@ type ResourceIdentity struct {
 	TenantID *string
 }
 
+// SelectorClassification provides polymorphic access to related types.
+// Call the interface's GetSelector() method to access the common type.
+// Use a type switch to determine the concrete type.  The possible types are:
+// - *ListSelector, *QuerySelector, *Selector
+type SelectorClassification interface {
+	// GetSelector returns the Selector content of the underlying type.
+	GetSelector() *Selector
+}
+
 // Selector - Model that represents a selector in the Experiment resource.
 type Selector struct {
 	// REQUIRED; String of the selector ID.
 	ID *string
 
-	// REQUIRED; List of Target references.
-	Targets []*TargetReference
-
 	// REQUIRED; Enum of the selector type.
 	Type *SelectorType
+
+	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
+	AdditionalProperties map[string]any
+
+	// Model that represents available filter types that can be applied to a targets list.
+	Filter FilterClassification
+}
+
+// GetSelector implements the SelectorClassification interface for type Selector.
+func (s *Selector) GetSelector() *Selector { return s }
+
+// SimpleFilter - Model that represents a simple target filter.
+type SimpleFilter struct {
+	// REQUIRED; Enum that discriminates between filter types. Currently only Simple type is supported.
+	Type *FilterType
+
+	// Model that represents the Simple filter parameters.
+	Parameters *SimpleFilterParameters
+}
+
+// GetFilter implements the FilterClassification interface for type SimpleFilter.
+func (s *SimpleFilter) GetFilter() *Filter {
+	return &Filter{
+		Type: s.Type,
+	}
+}
+
+// SimpleFilterParameters - Model that represents the Simple filter parameters.
+type SimpleFilterParameters struct {
+	// List of Azure availability zones to filter targets by.
+	Zones []*string
 }
 
 // Step - Model that represents a step in the Experiment resource.
@@ -743,9 +904,8 @@ type TargetReference struct {
 	// REQUIRED; String of the resource ID of a Target resource.
 	ID *string
 
-	// CONSTANT; Enum of the Target reference type.
-	// Field has constant value "ChaosTarget", any specified value is ignored.
-	Type *string
+	// REQUIRED; Enum of the Target reference type.
+	Type *TargetReferenceType
 }
 
 // TargetType - Model that represents a Target Type resource.